Backend/requirements.txt
python-dotenv
gradio
langchain-openai
langchain-core
langchain-community
azure-cosmos
langgraph